ID3 Decision Tree Implementation
This is a ground-up implementation of the ID3 algorithm for decision
tree learning. informationFunctions.py contains functions for
calculation entropy and information gain based on features, tree.py
contains a Node class for constructing the tree and functions for
testing the tree, main.py contains the ID3 algorithm and various
auxiliary functions, /Learning/ contains the training and testing
examples. To run, simply execute main.py.